requirements.txt
colorlog==6.9.0
homeassistant==2025.1.0
hassil>=1.7.0
pip>=21.0,<24.4
ruff==0.8.4
markdown-it-py==3.0.0
mdit-plain==1.0.1
